Replace cloud animation with css animations rather than transitions
These have the benefit of being infitely loopable, removing
the need to add/remove clouds when they reach the right-hand edge
and all the associated event handling.

They are also easier to set up, as we don't have to wait for the
element to be in the DOM to apply the animation.

However, this does mean that the same set of clouds continually
floats across the background, unlike the old approach. But because
the animation durations are randomised, there's still enough
change for it to continue to look interesting.