Wait before resizing to get better measurements
I noticed some issues with the app failing to fill the window
on Android, when it was rotated and it resized itself.

By putting some setTimeout() calls around the scale() function,
we should be able to get more accurate measurements, which should
reduce occurrences of this.

However, I still haven't been able to completely get rid of
these resizing artefacts.