Add swapper for torso
NB this doesn't animate correctly when the game is lost and the
man falls, as the torso and head are left in mid-air.

The code for managing the animations is such a mess that I'm
abandoning this. It's going to need a full rewrite of the
animation code to get it right.